tdf#103322 GlyphItem::m_nXOffset change from int to double
GlyphItem::m_nXOffset is one of the variables used to position a glyph.
To support sub pixel glyph positioning GlyphItem::m_nXOffset needs to be a
double.

Although GlyphItem::m_nXOffset is now a double we are still using it to store
the same integer values, as the rest of the layout code is still using integer
values.
But later on we will be using GlyphItem::m_nXOffset to store floating point
values.

+++ vcl/inc/impglyphitem.hxx
@@ -38,11 +38,11 @@ struct VCL_DLLPUBLIC GlyphItem
     int m_nCharPos; // index in string
     int m_nFlags;
     int m_nNewWidth; // width after adjustments
-    int m_nXOffset;
+    double m_nXOffset;
     Point m_aLinearPos; // absolute position of non rotated string
 
     GlyphItem(int nCharPos, int nCharCount, sal_GlyphId aGlyphId, const Point& rLinearPos,
-              long nFlags, int nOrigWidth, int nXOffset, LogicalFontInstance* pFontInstance)
+              long nFlags, int nOrigWidth, double nXOffset, LogicalFontInstance* pFontInstance)
         : m_aGlyphId(aGlyphId)
         , m_nCharCount(nCharCount)
         , m_nOrigWidth(nOrigWidth)
